Test cases for:
Fix regression caused by previous commits for CASTOR-385 by moving the name-collision check
added in CASTOR-385 to the class that actually uses the name.  Thus, the check is against
the name **after processing by the binding file** and the check is only on elements/types/etc
that will cause a source file to be written.